+++ backend/src/main/java/com/festago/presentation/AdminController.java
@@ -11,7 +11,6 @@ import com.festago.auth.dto.RootAdminInitializeRequest;
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.InternalServerException;
-import com.festago.common.exception.UnauthorizedException;
 import com.festago.festival.application.FestivalService;
 import com.festago.festival.dto.FestivalCreateRequest;
 import com.festago.festival.dto.FestivalResponse;
@@ -22,7 +21,6 @@ import com.festago.ticket.application.TicketService;
 import com.festago.ticket.dto.TicketCreateRequest;
 import com.festago.ticket.dto.TicketCreateResponse;
 import io.swagger.v3.oas.annotations.Hidden;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
@@ -31,16 +29,11 @@ import org.springframework.boot.info.BuildProperties;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.View;
-import org.springframework.web.servlet.view.InternalResourceView;
-import org.springframework.web.servlet.view.RedirectView;
 
 @RestController
 @RequestMapping("/admin")
@@ -84,16 +77,6 @@ public class AdminController {
         TicketCreateResponse response = ticketService.create(request);
         return ResponseEntity.ok()
             .body(response);
-    }
-
-    @GetMapping
-    public ModelAndView adminPage() {
-        return new ModelAndView("admin/admin-page");
-    }
-
-    @GetMapping("/login")
-    public ModelAndView loginPage() {
-        return new ModelAndView("admin/login");
     }
 
     @PostMapping("/login")
@@ -147,25 +130,11 @@ public class AdminController {
             .build();
     }
 
-    @GetMapping("/signup")
-    public ModelAndView signupPage() {
-        return new ModelAndView("admin/signup");
-    }
-
     @PostMapping("/signup")
     public ResponseEntity<AdminSignupResponse> signupAdminAccount(@RequestBody @Valid AdminSignupRequest request,
                                                                   @Admin Long adminId) {
         AdminSignupResponse response = adminAuthService.signup(adminId, request);
         return ResponseEntity.ok()
             .body(response);
-    }
-
-    @ExceptionHandler(UnauthorizedException.class)
-    public View handle(UnauthorizedException e, HttpServletResponse response) {
-        if (e.getErrorCode() == ErrorCode.EXPIRED_AUTH_TOKEN) {
-            return new RedirectView("/admin/login");
-        }
-        response.setStatus(HttpServletResponse.SC_NOT_FOUND);
-        return new InternalResourceView("/error/404");
     }
 }
